prepush lint
