+++ package.json
@@ -7,7 +7,8 @@
     "dev:main": "turbo run dev --filter=!./examples/*",
     "format": "prettier --write \"**/*.{ts,tsx,md}\"",
     "lint": "turbo run lint",
-    "sort-package-jsons": "npx sort-package-json \"**/package.json\" \"!**/node_modules/**\" --check"
+    "sort-package-jsons": "npx sort-package-json \"**/package.json\" \"!**/node_modules/**\" --check",
+    "prepare": "husky install"
   },
   "dependencies": {
     "@changesets/cli": "^2.26.1"
@@ -15,10 +16,14 @@
   "devDependencies": {
     "eslint": "^8.42.0",
     "eslint-config-custom": "workspace:*",
+    "husky": "^8.0.3",
     "prettier": "^2.8.8",
     "syncpack": "^10.5.1",
     "turbo": "^1.10.1",
     "typescript": "^5.1.3"
   },
-  "packageManager": "pnpm@7.15.0"
+  "packageManager": "pnpm@7.15.0",
+  "husky": {
+    "pre-push": "pnpm lint"
+  }
 }
